Update build
This commit updates the build to pass on the latest github runners. To
this end we:
* update the github workflows to use the latest versions of all actions
* enable the schnorr signature and the recovery module in the
installation of the `secp256k1` library on all platforms to resolve any
issues with missing symbols
* update the stackage resolver to version lts-22.6 - the macOS runner
has issues building on GHC 9.0.X
* did not upgrade but added lower bounds to the `secp256k1-haskell`
dependency of `< 1` as presumably we plan to move to the new
`libsecp256k1` library
* small changes `Bitcoin.Transaction.Builder` and `Bitcoin.Util` to
make the code compile with the latest version of `bytestring` and
remove the reliance on a missing (likely orphan) instances of
`Alternative` for `Either a`.

Signed-off-by: Tochi Obudulu <tochicool@gmail.com>